feat: added controller for pregnany db

+++ packages/optimise-core/src/utils/db-handler.js
@@ -5,10 +5,9 @@ import fs from 'fs';
 
 // Current level of the DB
 // This field is to be updated with subsequent versions of the DB
-const CURRENT_VERSION = 14;
+const CURRENT_VERSION = 15;
 
 export async function migrate() {
-
     // Verify the OPT_KV configuration table exists
     const isIntitialized = await dbcon().schema.hasTable('OPT_KV');
     let stepVersion = 0;
@@ -24,22 +23,30 @@ export async function migrate() {
             key: 'CURRENT_VERSION',
             value: '0',
             created_at: dbcon().fn.now(),
-            updated_at: dbcon().fn.now()
+            updated_at: dbcon().fn.now(),
         });
     } else {
         // Otherwise fetch the CURRENT_VERSION
-        let stepVersionResult = await dbcon()('OPT_KV').where({
-            key: 'CURRENT_VERSION'
-        }).select('value');
+        let stepVersionResult = await dbcon()('OPT_KV')
+            .where({
+                key: 'CURRENT_VERSION',
+            })
+            .select('value');
 
-        if (stepVersionResult.length === 1 && stepVersionResult[0].value !== undefined)
+        if (
+            stepVersionResult.length === 1 &&
+            stepVersionResult[0].value !== undefined
+        )
             stepVersion = parseInt(stepVersionResult[0].value) || 0;
     }
 
     if (stepVersion !== CURRENT_VERSION) {
-
         if (CURRENT_VERSION < stepVersion)
-            return Promise.reject(new Error('The existing database was created with a newer version of Optimise ! Please upgrade before using Optimise !'));
+            return Promise.reject(
+                new Error(
+                    'The existing database was created with a newer version of Optimise ! Please upgrade before using Optimise !'
+                )
+            );
 
         // For every table file launch the update for sequential version up
         while (stepVersion < CURRENT_VERSION) {
@@ -55,12 +62,14 @@ export async function migrate() {
         }
 
         // Finally set the CURRENT_VERSION to the current level
-        await dbcon()('OPT_KV').where({
-            key: 'CURRENT_VERSION'
-        }).update({
-            value: `${stepVersion}`,
-            updated_at: dbcon().fn.now()
-        });
+        await dbcon()('OPT_KV')
+            .where({
+                key: 'CURRENT_VERSION',
+            })
+            .update({
+                value: `${stepVersion}`,
+                updated_at: dbcon().fn.now(),
+            });
     }
 
     return dbcon;
@@ -69,11 +78,11 @@ export async function migrate() {
 export function erase() {
     return new Promise((resolve, reject) => {
         if (process.env.NODE_ENV !== 'test') {
-            if (process.env.NODE_ENV === 'development') console.log('Removing database file ...');
+            if (process.env.NODE_ENV === 'development')
+                console.log('Removing database file ...');
             let filename = dbcon().client.config.connection.filename;
             try {
-                if (fs.existsSync(filename))
-                    fs.unlinkSync(filename);
+                if (fs.existsSync(filename)) fs.unlinkSync(filename);
             } catch (err) {
                 return reject(err);
             }
@@ -84,5 +93,5 @@ export function erase() {
 
 export default {
     migrate,
-    erase
+    erase,
 };
